# PoS improvements