--- a/crates/subspace-proof-of-space/Cargo.toml
+++ b/crates/subspace-proof-of-space/Cargo.toml
@@ -18,9 +18,12 @@ bench = false
 [dependencies]
 chacha20 = { version = "0.9.1", default-features = false }
 derive_more = "0.99.17"
+parking_lot = { version = "0.12.2", optional = true }
 rayon = { version = "1.10.0", optional = true }
 seq-macro = "0.3.5"
 sha2 = { version = "0.10.7", default-features = false }
+# Replacement for `parking_lot` in `no_std` environment
+spin = "0.9.7"
 subspace-core-primitives = { version = "0.1.0", path = "../subspace-core-primitives", default-features = false }
 
 [dev-dependencies]
@@ -38,6 +41,8 @@ harness = false
 default = ["std"]
 std = [
     "chacha20/std",
+    # In no-std environment we use `spin`
+    "parking_lot",
     "sha2/std",
     "subspace-core-primitives/std",
 ]