Supranational repored that out code didn't find all proofs/qualities, which while didn't actually matter for Subspace, was strictly speaking slightly wrong, I fixed that in the first commit.

Then I found two more ways to make table generation faster, which cumulatively resulted in a nice performance improvement.

Before:
```
chia/table/parallel/1x  time:   [192.64 ms 197.94 ms 203.28 ms]
chia/table/parallel/8x  time:   [949.34 ms 953.08 ms 957.16 ms]
```

After:
```
chia/table/parallel/1x  time:   [149.95 ms 152.86 ms 155.13 ms]
chia/table/parallel/8x  time:   [913.09 ms 925.22 ms 938.36 ms]
```

I had nicer API for `LeftTables` with methods and two more data structures, but it was not clear how to convince Rust to agree that lifetimes are correct, so I had to resort to memory slices in `find_matches`. Future borrow checked in Rust should help cleaning up the code a bit.

I tried to apply the same trick with not collecting buckets as in third commit for single-threaded implementation, but I think due to code cache it was slightly slower to the same, but more code and harder to read, so I didn't commit those changes.

I also experimented with data alignment and a few other things, but none of them resulted in conclusive consistent performance improvement.

There might be more performance optimizations, but for now I exhausted ideas and not looking forward to spend hours reading x86-64 assembly looking for opportunities.

### Code contributor checklist:
* [x] I have read, understood and followed [contributing guide](https://github.com/subspace/subspace/blob/main/CONTRIBUTING.md)