@dariolina 2024-05-20
Agree that now it looks more correct, but it makes proof generation regress 3-10% on my machine. Do you see similar performance?