@nazar-pc 2024-05-20
I see a small performance decrease in case of missing proof, but only because it is already doing 150M of them per second.

Before:
```
chia/proof/missing      time:   [33.089 ns 33.166 ns 33.283 ns]
chia/proof/present      time:   [365.81 ns 366.21 ns 366.46 ns]
```

After:
```
chia/proof/missing      time:   [35.223 ns 35.273 ns 35.333 ns]
chia/proof/present      time:   [366.60 ns 366.97 ns 367.35 ns]
```

Seems acceptable to me given how little compute it requires overall.